Extract shared locking helpers from serial_task
Moves get_unique_key, run_with_lock and
CouldNotAcquireLockError out of corehq/apps/celery/serial.py into a new
corehq/apps/celery/locking.py module so they can be reused by an
upcoming concurrent_task decorator. No behavior change.
