+++ corehq/apps/celery/serial.py
@@ -1,6 +1,10 @@
-import inspect
 from functools import wraps
 
+from corehq.apps.celery.locking import (
+    CouldNotAcquireLockError,
+    get_unique_key,
+    run_with_lock,
+)
 from corehq.apps.celery.shared_task import task
 
 
@@ -53,44 +57,12 @@ def serial_task(
         )
         @wraps(fn)
         def _inner(self, *args, **kwargs):
-            key = _get_unique_key(unique_key, fn, *args, **kwargs)
+            key = get_unique_key(unique_key, fn, *args, **kwargs)
             try:
-                return _run_with_lock(key, fn, timeout, *args, **kwargs)
+                return run_with_lock(key, fn, timeout, *args, **kwargs)
             except CouldNotAcquireLockError as e:
                 self.retry(exc=e)
 
         return _inner
 
     return decorator
-
-
-def _run_with_lock(key, fn, timeout, *args, **kwargs):
-    from dimagi.utils.couch import get_redis_lock, release_lock
-
-    lock = get_redis_lock(key, timeout=timeout, name=fn.__name__)
-    if lock.acquire(blocking=False):
-        try:
-            return fn(*args, **kwargs)
-        finally:
-            release_lock(lock, True)
-    else:
-        msg = f"Could not acquire lock '{key}' for task '{fn.__name__}'"
-        raise CouldNotAcquireLockError(msg)
-
-
-def _get_unique_key(format_str, fn, *args, **kwargs):
-    """
-    Builds a unique key from the function name and ``format_str``.
-
-    Binds args and kwargs to the function's signature which enables
-    referencing any arg or kwarg by name in ``format_str``
-
-    See corehq.apps.celery.tests.test_get_unique_key for more details.
-    """
-    bound = inspect.signature(fn).bind(*args, **kwargs)
-    bound.apply_defaults()  # ensures bound.arguments includes default values
-    return f"{fn.__name__}-{format_str.format(**bound.arguments)}"
-
-
-class CouldNotAcquireLockError(Exception):
-    """Used when a serial_task is unable to obtain lock to run"""
